refactor: extract command dispatch into separate module
- Create command_dispatch.hpp with CommandResult, CommandHandler, CommandTable type aliases
- Move register_command(), register_commands(), dispatch_command() to command_dispatch.cpp
- app_base.hpp includes command_dispatch.hpp for the shared types
- app_base.cpp is now focused on application lifecycle, no command handler logic
