+++ include/gseurat/engine/app_base.hpp
@@ -8,6 +8,7 @@
 #include "gseurat/engine/control_server.hpp"
 #endif
 #include "gseurat/engine/collision_gen.hpp"
+#include "gseurat/engine/command_dispatch.hpp"
 #include "gseurat/engine/coordinate.hpp"
 #include "gseurat/engine/gaussian_cloud.hpp"
 #include "gseurat/engine/day_night_system.hpp"
@@ -37,10 +38,7 @@
 
 #include <chrono>
 #include <cstdint>
-#include <expected>
 #include <functional>
-#include <string>
-#include <unordered_map>
 #include <vector>
 
 namespace gseurat {
@@ -246,12 +244,6 @@ protected:
     float play_time_ = 0.0f;
 
     // ── Command dispatch (C++23 Command Pattern) ──
-    // CommandResult: std::expected — success carries the JSON response,
-    // failure carries an error message string.
-    using CommandResult = std::expected<nlohmann::json, std::string>;
-    // std::move_only_function when compiler supports P1288R9; std::function for now.
-    using CommandHandler = std::function<CommandResult(const nlohmann::json&)>;
-
     void register_command(std::string name, CommandHandler handler);
     virtual void register_commands();
     virtual void dispatch_command(const nlohmann::json& cmd, nlohmann::json& response);
@@ -261,7 +253,7 @@ protected:
 #ifndef _WIN32
     ControlServer control_server_;
 #endif
-    std::unordered_map<std::string, CommandHandler> command_handlers_;
+    CommandTable command_handlers_;
 
     // Terrain PLY AABB (for grid→world coordinate conversion)
     AABB terrain_aabb_;
